third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc-error.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<body>
<script>
    document.body.innerHTML += '<img id="error" srcset="http://127.0.0.1:8000/resources/slow-image.php?name=not_there.png&mimeType=image/png&sleep=100">';
    jsTestIsAsync = true;
    var img_err = window.document.getElementById("error");
    img_err.onerror = function() {
        shouldBe('img_err.currentSrc', '"http://127.0.0.1:8000/resources/slow-image.php?name=not_there.png&mimeType=image/png&sleep=100"');
    };

    window.onload = function() {
        finishJSTest();
    };
</script>
</body>
