third_party/blink/web_tests/http/tests/images/image-currentsrc-error-expected.txt

PASS img_err.currentSrc is "http://127.0.0.1:8000/resources/slow-image.php?name=not_there.png&mimeType=image/png&sleep=100"
PASS successfullyParsed is true

TEST COMPLETE
